Incorporated review feedback, refactored code to be more generic
